feat: implement PersonsContext
