+++ src/contexts/PersonsContext.tsx
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { Person } from '../interfaces/SWApi';
+
+export const PersonsContext = createContext<Person[]>([]);
