+++ src/hooks/useSearch.ts
@@ -2,10 +2,5 @@ import { useContext } from 'react';
 import { SearchContext } from '../contexts/SearchContext';
 
 export function useSearch() {
-  const search = useContext(SearchContext);
-
-  if (search == null) {
-    throw new Error('useSearch has to be used within <SearchContext.Provider>');
-  }
-  return search;
+  return useContext(SearchContext);
 }
